create new api and transaction

+++ src/main/java/com/account/dashboard/serviceImpl/VoucherServiceImpl.java
@@ -311,9 +311,9 @@ public class VoucherServiceImpl implements VoucherService{
 		String dateString = dateFormat.format(date);
 		List<Voucher>voucherList=voucherRepository.findByIdInBetween(startDate,dateString);
 
-		long totalCredit=0;
-		long totalDebit=0;
-		long totalAmount=0;
+		double totalCredit=0;
+		double totalDebit=0;
+		double totalAmount=0;
 		for(Voucher v:voucherList) {
 			Map<String,Object>map = new HashMap<>();
 			map.put("id", v.getId());
@@ -334,14 +334,14 @@ public class VoucherServiceImpl implements VoucherService{
 			res.add(map);
 
 			if(v.isCreditDebit()) {
-				long debitAmount =Long.valueOf(v.getDebitAmount()!=null?v.getDebitAmount():"0");
-				long creditAmount =Long.valueOf(v.getCreditAmount()!=null?v.getCreditAmount():"0");
+				double debitAmount =Double.valueOf(v.getDebitAmount()!=null?v.getDebitAmount():"0");
+				double creditAmount =Double.valueOf(v.getCreditAmount()!=null?v.getCreditAmount():"0");
 				totalCredit=totalCredit+creditAmount;
 				totalDebit=totalDebit+debitAmount;
 
 				totalAmount=totalAmount-debitAmount+creditAmount;
 			}else {
-				long debitAmount =Long.valueOf(v.getDebitAmount()!=null?v.getDebitAmount():"0");
+				double debitAmount =Double.valueOf(v.getDebitAmount()!=null?v.getDebitAmount():"0");
 				totalDebit=totalDebit+debitAmount;
 				totalAmount=totalAmount-debitAmount;
